+++ src/lib.rs
@@ -28,9 +28,11 @@
 pub mod alignment;
 pub mod db;
 pub mod formats;
+pub mod path_index;
 pub mod subgraph;
 pub mod utils;
 
 pub use alignment::Alignment;
 pub use db::{GBZBase, GAFBase, GBZPath, GBZRecord, GraphInterface};
-pub use subgraph::{Subgraph, SubgraphQuery, HaplotypeOutput, PathIndex};
+pub use path_index::PathIndex;
+pub use subgraph::{Subgraph, SubgraphQuery, HaplotypeOutput};
