+++ src/subgraph.rs
@@ -4,11 +4,11 @@
 //! The subgraph contains all nodes within a given context and all edges between them.
 //! All other paths within the subgraph can also be extracted, but they will not have any true metadata associated with them.
 
-use crate::{GBZRecord, GBZPath, GraphInterface};
+use crate::{GBZRecord, GBZPath, GraphInterface, PathIndex};
 use crate::formats::{self, WalkMetadata, JSONValue};
 
 use std::cmp::Reverse;
-use std::collections::{BinaryHeap, BTreeMap, BTreeSet, HashMap};
+use std::collections::{BinaryHeap, BTreeMap, BTreeSet};
 use std::fmt::Display;
 use std::io::{self, Write};
 use std::ops::Range;
@@ -18,9 +18,6 @@ use gbwt::ENDMARKER;
 use gbwt::{GBZ, GBWT, GraphPosition, Orientation, Pos, FullPathName};
 
 use gbwt::{algorithms, support};
-
-use simple_sds::sparse_vector::{SparseVector, SparseBuilder};
-use simple_sds::ops::PredSucc;
 
 #[cfg(test)]
 mod tests;
@@ -203,181 +200,22 @@ impl Display for SubgraphQuery {
 
 //-----------------------------------------------------------------------------
 
-/// An index for random access to reference and generic paths in a GBZ graph.
-///
-/// Indexed paths are identified by their offsets in the index.
-/// The offsets range from 0 to `path_count() - 1`.
-///
-/// The combination of [`GBZ`] and [`PathIndex`] is functionally similar to [`GraphInterface`] but tens of times faster.
-/// An in-memory graph is better for batch operations, where the loading time is a negligible fraction of the total time.
-/// The database is better for interactive applications, where the user works with relatively small subgraphs.
-/// For a human graph, the database should be faster than the in-memory graph for subgraphs of up to 1 Mbp.
-///
-/// # Examples
-///
-/// ```
-/// use gbz_base::PathIndex;
-/// use gbwt::{GBZ, FullPathName, Pos, REF_SAMPLE};
-/// use gbwt::support;
-/// use simple_sds::serialize;
-///
-/// let filename = support::get_test_data("example.gbz");
-/// let graph: GBZ = serialize::load_from(&filename).unwrap();
-///
-/// // Create a path index with 3 bp intervals.
-/// let path_index = PathIndex::new(&graph, 3, false);
-/// assert!(path_index.is_ok());
-/// let path_index = path_index.unwrap();
-///
-/// // We have two components with one generic path in each.
-/// assert_eq!(path_index.path_count(), 2);
-/// assert_eq!(path_index.path_length(0), Some(5));
-/// assert_eq!(path_index.path_length(1), Some(4));
-///
-/// // Consider the generic path in component A.
-/// let path_name = FullPathName::generic("A");
-/// let index_offset = path_index.find_path(&graph, &path_name);
-/// assert_eq!(index_offset, Some(0));
-/// let index_offset = index_offset.unwrap();
-///
-/// // There should be two indexed positions for the path.
-/// let first_sample = path_index.indexed_position(index_offset, 2);
-/// assert_eq!(first_sample, Some((0, Pos::new(22, 0))));
-/// let second_sample = path_index.indexed_position(index_offset, 5);
-/// assert_eq!(second_sample, Some((3, Pos::new(30, 0))));
-/// let next_sample = path_index.indexed_position(index_offset, 100);
-/// assert_eq!(next_sample, second_sample);
-/// ```
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct PathIndex {
-    // Maps path identifiers to index offsets.
-    path_to_offset: HashMap<usize, usize>,
+// FIXME: Tests for the new functionality.
 
-    // Maps index offsets to path identifiers.
-    offset_to_path: HashMap<usize, usize>,
-
-    // Sequence lengths for each path in bp.
-    path_lengths: Vec<usize>,
-
-    // Indexed sequence positions for each path.
-    sequence_positions: Vec<SparseVector>,
-
-    // GBWT positions corresponding to the indexed sequence positions.
-    gbwt_positions: Vec<Vec<Pos>>,
-}
-
-impl PathIndex {
-    /// Creates a new path index for the given GBZ graph.
-    ///
-    /// The index is built for all reference and generic paths.
-    ///
-    /// # Arguments
-    ///
-    /// * `graph`: A GBZ graph.
-    /// * `interval`: Approximate distance between indexed positions (in bp).
-    /// * `verbose`: Print progress information to stderr.
-    pub fn new(graph: &GBZ, interval: usize, verbose: bool) -> Result<Self, String> {
-        if verbose {
-            eprintln!("Building path index");
-        }
-
-        // NOTE: For fragmented haplotypes, the reference positions are relative
-        // to the start of the fragment.
-        let reference_paths = graph.reference_positions(interval, verbose);
-        if reference_paths.is_empty() {
-            return Err(String::from("No reference paths to index"));
-        }
-
-        let mut path_to_offset: HashMap<usize, usize> = HashMap::with_capacity(reference_paths.len());
-        let mut offset_to_path: HashMap<usize, usize> = HashMap::with_capacity(reference_paths.len());
-        let mut path_lengths: Vec<usize> = Vec::with_capacity(reference_paths.len());
-        let mut sequence_positions = Vec::with_capacity(reference_paths.len());
-        let mut gbwt_positions = Vec::with_capacity(reference_paths.len());
-        for ref_path in reference_paths.iter() {
-            path_to_offset.insert(ref_path.id, path_to_offset.len());
-            offset_to_path.insert(offset_to_path.len(), ref_path.id);
-            path_lengths.push(ref_path.len);
-            let mut sequence = SparseBuilder::new(ref_path.len, ref_path.positions.len()).map_err(
-                String::from
-            )?;
-            let mut gbwt = Vec::with_capacity(ref_path.positions.len());
-            for (sequence_pos, gbwt_pos) in ref_path.positions.iter() {
-                sequence.set(*sequence_pos);
-                gbwt.push(*gbwt_pos);
-            }
-            sequence_positions.push(SparseVector::try_from(sequence).map_err(String::from)?);
-            gbwt_positions.push(gbwt);
-        }
-
-        Ok(PathIndex { path_to_offset, offset_to_path, path_lengths, sequence_positions, gbwt_positions })
-    }
-
-    /// Returns the number of indexed paths.
-    pub fn path_count(&self) -> usize {
-        self.sequence_positions.len()
-    }
-
-    /// Returns the index offset for the indexed path with the given identifier.
-    ///
-    /// Returns [`None`] if the path does not exist or it has not been indexed.
-    pub fn path_to_offset(&self, path_id: usize) -> Option<usize> {
-        self.path_to_offset.get(&path_id).cloned()
-    }
-
-    /// Returns the path identifier for the indexed path with the given index offset.
-    ///
-    /// Returns [`None`] if the path does not exist.
-    pub fn offset_to_path(&self, index_offset: usize) -> Option<usize> {
-        self.offset_to_path.get(&index_offset).cloned()
-    }
-
-    /// Returns the index offset for the path with the given metadata.
-    ///
-    /// Returns [`None`] if the path does not exist or it has not been indexed.
-    pub fn find_path(&self, graph: &GBZ, path_name: &FullPathName) -> Option<usize> {
-        let metadata = graph.metadata()?;
-        let path_id = metadata.find_path(path_name)?;
-        self.path_to_offset(path_id)
-    }
-
-    /// Returns the length of the indexed path with the given index offset.
-    ///
-    /// Returns [`None`] if the path does not exist.
-    pub fn path_length(&self, index_offset: usize) -> Option<usize> {
-        self.path_lengths.get(index_offset).cloned()
-    }
-
-    /// Returns the last indexed position at or before `offset` on the path with name `path_name`.
-    ///
-    /// The return value consists of a sequence offset and a GBWT position.
-    /// Returns [`None`] if the path does not exist or it has not been indexed.
-    /// This is similar to [`GraphInterface::find_path`] followed by [`GraphInterface::indexed_position`].
-    ///
-    /// # Arguments
-    ///
-    /// * `index_offset`: Offset of the path in the index.
-    /// * `query_offset`: Sequence position in the path (in bp).
-    pub fn indexed_position(&self, index_offset: usize, query_offset: usize) -> Option<(usize, Pos)> {
-        let mut iter = self.sequence_positions[index_offset].predecessor(query_offset);
-        if let Some((sample_offset, sequence_offset)) = iter.next() {
-            let gbwt_pos = self.gbwt_positions[index_offset][sample_offset];
-            Some((sequence_offset, gbwt_pos))
-        } else {
-            None
-        }
-    }
-}
-
-//-----------------------------------------------------------------------------
-
-// FIXME: Tests and examples for the new functionality.
-
-// FIXME: Document the intended use of this struct, for example as a sliding window over a reference path.
+// FIXME: example that replicates from_gbz with a path offset using lower-level functions.
 /// A subgraph based on a context around a graph position.
 ///
 /// The position can be based on a path offset, a path interval, or a node identifier.
 /// The path used for extracting the subgraph becomes the reference path for it.
 /// Non-reference haplotypes do not have any metadata associated with them, as we cannot determine the identifier of a path from its GBWT position efficiently.
+///
+/// This struct is intended to be used as a sliding window in the graph.
+/// Individual nodes can be added and removed with the [`Subgraph::add_node`] and [`Subgraph::remove_node`] methods.
+/// The subgraph can be updated to a new context with the [`Subgraph::around_position`] and [`Subgraph::around_node`] methods.
+/// These methods reuse existing records when possible.
+/// Changes to the subgraph do not preserve paths; they must be re-extracted with the [`Subgraph::extract_paths`] method.
+/// [`Subgraph::from_gbz`] and [`Subgraph::from_db`] are convenience methods for extracting a subgraph using a [`SubgraphQuery`].
+/// The subgraph can be converted to GFA or JSON with the [`Subgraph::write_gfa`] and [`Subgraph::write_json`] methods.
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct Subgraph {
     // Node records for the subgraph.
@@ -662,20 +500,12 @@ impl Subgraph {
             let path_index = path_index.ok_or(
                 String::from("Path index is required for path-based queries")
             )?;
-            let path_name = query.path_name_with_offset();
-            let ref_path = GBZPath::with_name(graph, &path_name).ok_or(
-                format!("Cannot find a path covering {}", path_name)
-            )?;
-            // Transform the offset relative to the haplotype to the offset relative to the path.
-            let query_offset = path_name.fragment - ref_path.name.fragment;
-
-            // Now we can build the subgraph.
-            let ref_pos = query_position_from_gbz(graph, path_index, &ref_path, query_offset)?;
-            let reference_path = Some((ref_pos, ref_path.name));
-            self.around_position(ref_pos.graph_pos, query.context, &mut |handle| {
+            let query_pos = query.path_name_with_offset();
+            let reference_path = reference_position_from_gbz(graph, path_index, query_pos)?;
+            self.around_position(reference_path.0.graph_pos, query.context, &mut |handle| {
                 GBZRecord::from_gbz(graph, handle).ok_or(format!("The graph does not contain handle {}", handle))
             })?;
-            self.extract_paths(reference_path, query.output())?;
+            self.extract_paths(Some(reference_path), query.output())?;
         } else if query.is_node_based() {
             if query.output() == HaplotypeOutput::ReferenceOnly {
                 return Err(String::from("Cannot output a reference path in a node-based query"));
@@ -741,23 +571,13 @@ impl Subgraph {
     pub fn from_db(&mut self, graph: &mut GraphInterface, query: &SubgraphQuery) -> Result<(), String> {
         // FIXME: Do interval-based properly.
         if query.is_path_based() {
-            let path_name = query.path_name_with_offset();
-            let ref_path = graph.find_path(&path_name)?;
-            let ref_path = ref_path.ok_or(format!("Cannot find a path covering {}", path_name))?;
-            if !ref_path.is_indexed {
-                return Err(format!("Path {} has not been indexed for random access", ref_path.name));
-            }
-            // Transform the offset relative to the haplotype to the offset relative to the path.
-            let query_offset = path_name.fragment - ref_path.name.fragment;
-
-            // Now we can build the subgraph.
-            let ref_pos = query_position_from_db(graph, &ref_path, query_offset)?;
-            let reference_path = Some((ref_pos, ref_path.name));
-            self.around_position(ref_pos.graph_pos, query.context, &mut |handle| {
+            let query_pos = query.path_name_with_offset();
+            let reference_path = reference_position_from_db(graph, query_pos)?;
+            self.around_position(reference_path.0.graph_pos, query.context, &mut |handle| {
                 let record = graph.get_record(handle)?;
                 record.ok_or(format!("The graph does not contain handle {}", handle))
             })?;
-            self.extract_paths(reference_path, query.output())?;
+            self.extract_paths(Some(reference_path), query.output())?;
         } else if query.is_node_based() {
             if query.output() == HaplotypeOutput::ReferenceOnly {
                 return Err(String::from("Cannot output a reference path in a node-based query"));
@@ -789,6 +609,7 @@ impl Subgraph {
         }
     }
 
+    // FIXME: Example using add_node and remove_node.
     /// Extracts all paths in the subgraph.
     ///
     /// Clears the current paths in the subgraph.
@@ -1032,8 +853,8 @@ impl Subgraph {
         self.paths.len()
     }
 
-    /// Clears all path information from the subgraph.
-    pub fn clear_paths(&mut self) {
+    // Clears all path information from the subgraph.
+    fn clear_paths(&mut self) {
         self.paths.clear();
         self.ref_id = None;
         self.ref_path = None;
@@ -1377,8 +1198,27 @@ impl Subgraph {
 
 //-----------------------------------------------------------------------------
 
-// Returns the graph position and the GBWT position for the given offset.
-fn query_position_from_db(graph: &mut GraphInterface, path: &GBZPath, query_offset: usize) -> Result<PathPosition, String> {
+// FIXME: This should be a method that updates the subgraph.
+// FIXME: example
+/// Returns the path position for the haplotype offset represented by the query position.
+///
+/// `query_pos.fragment` is used as an offset in the haplotype.
+/// The return value consists of the position and metadata for the path covering the position.
+///
+/// # Errors
+///
+/// Returns an error if database operations fail.
+/// Returns an error if there is no path covering the given position or the path has not been indexed for random access.
+pub fn reference_position_from_db(graph: &mut GraphInterface, query_pos: FullPathName) -> Result<(PathPosition, FullPathName), String> {
+    let path = graph.find_path(&query_pos)?;
+    let path = path.ok_or(format!("Cannot find a path covering {}", query_pos))?;
+    if !path.is_indexed {
+        return Err(format!("Path {} has not been indexed for random access", query_pos));
+    }
+    // Transform the offset relative to the haplotype to the offset relative to the path.
+    let query_offset = query_pos.fragment - path.name.fragment;
+
+    // Find an indexed position before the query position.
     let result = graph.indexed_position(path.handle, query_offset)?;
     let (mut path_offset, mut pos) = result.ok_or(
         format!("Path {} has not been indexed for random access", path.name())
@@ -1412,15 +1252,32 @@ fn query_position_from_db(graph: &mut GraphInterface, path: &GBZPath, query_offs
         format!("Path {} does not contain offset {}", path.name(), query_offset)
     )?;
     let gbwt_pos = gbwt_pos.unwrap();
-    Ok(PathPosition {
+
+    let path_position = PathPosition {
         seq_offset: query_offset,
         graph_pos,
         gbwt_pos,
-    })
+    };
+    Ok((path_position, path.name))
 }
 
-// Returns the graph position and the GBWT position for the given offset.
-fn query_position_from_gbz(graph: &GBZ, path_index: &PathIndex, path: &GBZPath, query_offset: usize) -> Result<PathPosition, String> {
+// FIXME: This should be a method that updates the subgraph.
+// FIXME: example
+/// Returns the path position for the haplotype offset represented by the query position.
+///
+/// `query_pos.fragment` is used as an offset in the haplotype.
+/// The return value consists of the position and metadata for the path covering the position.
+///
+/// # Errors
+///
+/// Returns an error if there is no path covering the given position or the path has not been indexed for random access.
+pub fn reference_position_from_gbz(graph: &GBZ, path_index: &PathIndex, query_pos: FullPathName) -> Result<(PathPosition, FullPathName), String> {
+    let path = GBZPath::with_name(graph, &query_pos).ok_or(
+        format!("Cannot find a path covering {}", query_pos)
+    )?;
+    // Transform the offset relative to the haplotype to the offset relative to the path.
+    let query_offset = query_pos.fragment - path.name.fragment;
+
     // Path id to an indexed position.
     let index_offset = path_index.path_to_offset(path.handle).ok_or(
         format!("Path {} has not been indexed for random access", path.name())
@@ -1454,11 +1311,13 @@ fn query_position_from_gbz(graph: &GBZ, path_index: &PathIndex, path: &GBZPath, 
         format!("Path {} does not contain offset {}", path.name(), query_offset)
     )?;
     let gbwt_pos = gbwt_pos.unwrap();
-    Ok(PathPosition {
+
+    let path_position = PathPosition {
         seq_offset: query_offset,
         graph_pos,
         gbwt_pos,
-    })
+    };
+    Ok((path_position, path.name))
 }
 
 //-----------------------------------------------------------------------------
@@ -1466,7 +1325,7 @@ fn query_position_from_gbz(graph: &GBZ, path_index: &PathIndex, path: &GBZPath, 
 /// A path position represented in multiple ways.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct PathPosition {
-    /// Offset in the path in bp.
+    /// Sequence offset in in bp.
     pub seq_offset: usize,
     /// Graph position corresponding to the offset.
     pub graph_pos: GraphPosition,
